input: misc: Avoid integer overflow and truncation
uinput_ioctl_handler returns a long value whereas the return
variable "ret" is declared as an int.
This may lead to integer overflow or truncation when a long value
is assigned to ret.
Fixed by declaring ret as a long variable.

Coverity ID 33459

Reviewed-on: http://git-master/r/1314873
(cherry picked from commit 345a5de346548435d9857b16784d7d4757fc83e9)
Reviewed-on: http://git-master/r/1493939
GVS: Gerrit_Virtual_Submit

+++ drivers/input/misc/uinput.c
@@ -790,7 +790,7 @@ static int uinput_str_to_user(void __user *dest, const char *str,
 static long uinput_ioctl_handler(struct file *file, unsigned int cmd,
 				 unsigned long arg, void __user *p)
 {
-	int			retval;
+	long			retval;
 	struct uinput_device	*udev = file->private_data;
 	struct uinput_ff_upload ff_up;
 	struct uinput_ff_erase  ff_erase;
